abstract the common tests into a test case
